+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## [Unreleased]
 ### Changed
 - Stopped tracking concatenated js. Deployment is handled by git hooks post-commit and post-merge.
+- Different respawn messages on touch and desktop.
 
 ## [0.11.0] - 2016-12-20
 ### Added